refactor: change schema

+++ server/src/services/UsersService.js
@@ -1,14 +1,15 @@
 const Error = require("../middlewares/Error");
 const DiskStorage = require("../providers/DiskStorage");
+const checkIfUserIsNotLogged = require("../utils/checkIfUserIsNotLogged");
 
-const { hash, compare } = require("bcrypt");
+const { hash } = require("bcrypt");
 
 class UsersService {
   constructor(usersRepository) {
     this.usersRepository = usersRepository;
   }
 
-  async create({ name, email, password, passwordConfirmation }) {
+  async create(name, email, password, passwordConfirmation) {
     if (!name || !email || !password || !passwordConfirmation) {
       throw new Error("Mandatory field's not informed");
     }
@@ -17,47 +18,29 @@ class UsersService {
       throw new Error("Typed password do not match");
     }
 
-    const userExists = await this.usersRepository.findByEmail({ email });
+    const emailExists = await this.usersRepository.findByEmail(email);
 
-    if (userExists) {
+    if (emailExists) {
       throw new Error("Email already used");
     }
 
     const hashedPassword = await hash(password, 8);
 
-    return await this.usersRepository.create({
-      name,
-      email,
-      password: hashedPassword,
-    });
+    return await this.usersRepository.create(name, email, hashedPassword);
   }
 
-  async update({ id, name, password, passwordConfirmation }) {
-    if (!name || !password || !passwordConfirmation) {
+  async updateImage(loggedUserId, image) {
+    const user = await this.usersRepository.findById(loggedUserId);
+
+    if (!loggedUserId || !user) {
+      throw new Error("User not logged");
+    }
+
+    if (!image) {
       throw new Error("Mandatory field's not informed");
     }
 
-    if (password !== passwordConfirmation) {
-      throw new Error("Typed password do not match");
-    }
-
-    const hashedPassword = await hash(password, 8);
-
-    return await this.usersRepository.update({
-      id,
-      name,
-      password: hashedPassword,
-    });
-  }
-
-  async updateImage({ id, image }) {
     const diskStorage = new DiskStorage();
-
-    const user = await this.usersRepository.findById({ id });
-
-    if (!user) {
-      throw new Error("Only authenticated users can update an avatar", 401);
-    }
 
     if (user.image) {
       await diskStorage.deleteFile(user.image);
@@ -67,11 +50,54 @@ class UsersService {
 
     user.image = filename;
 
-    const updatedUser = await this.usersRepository.updateImage({
-      id,
-      image: user.image,
-    });
+    const updatedUser = await this.usersRepository.updateImage(
+      loggedUserId,
+      user.image
+    );
     return updatedUser;
+  }
+
+  async update(loggedUserId, name, email, password, passwordConfirmation) {
+    checkIfUserIsNotLogged(this.usersRepository, loggedUserId);
+
+    if (!name || !email || !password || !passwordConfirmation) {
+      throw new Error("Mandatory field's not informed");
+    }
+
+    if (password !== passwordConfirmation) {
+      throw new Error("Typed password do not match");
+    }
+
+    const hashedPassword = await hash(password, 8);
+
+    return this.usersRepository.update(
+      loggedUserId,
+      name,
+      email,
+      hashedPassword
+    );
+  }
+
+  async delete(loggedUserId) {
+    checkIfUserIsNotLogged(this.usersRepository, loggedUserId);
+
+    return await this.usersRepository.delete(loggedUserId);
+  }
+
+  async findAll(loggedUserId) {
+    checkIfUserIsNotLogged(this.usersRepository, loggedUserId);
+
+    return await this.usersRepository.findAll();
+  }
+
+  async findById(loggedUserId, userIdToFind) {
+    checkIfUserIsNotLogged(this.usersRepository, loggedUserId);
+
+    if (!userIdToFind) {
+      throw new Error("Mandatory field's not informed");
+    }
+
+    return await this.usersRepository.findById(userIdToFind.id);
   }
 }
